fix(patina): report template child index keys
